matelog_backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-tinymce==5.0.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.4
tzdata==2025.2
whitenoise==6.11.0
